Clean up Claude's changes
