+++ machine/corpora/place_markers_usfm_update_block_handler.py
@@ -4,7 +4,7 @@ from typing import List, TypedDict, cast
 
 from ..translation.word_alignment_matrix import WordAlignmentMatrix
 from .segment_boundary_adjuster import SegmentBoundaryAdjuster
-from .update_usfm_parser_handler import UpdateUsfmMarkerBehavior
+from .update_usfm_behavior import UpdateUsfmMarkerBehavior
 from .usfm_token import UsfmToken, UsfmTokenType
 from .usfm_update_block import UsfmUpdateBlock
 from .usfm_update_block_element import UsfmUpdateBlockElement, UsfmUpdateBlockElementType
@@ -19,24 +19,6 @@ class PlaceMarkersAlignmentInfo(TypedDict):
     alignment: WordAlignmentMatrix
     paragraph_behavior: UpdateUsfmMarkerBehavior
     style_behavior: UpdateUsfmMarkerBehavior
-
-
-def _element_text(element: UsfmUpdateBlockElement) -> str:
-    # A TEXT element holds one token per row matched to the block, so a verse range matched by
-    # several rows has several tokens. Reading only the first drops the rest of the translation.
-    return "".join(t.to_usfm() for t in element.tokens)
-
-
-def _is_placeable(element: UsfmUpdateBlockElement, alignment_info: PlaceMarkersAlignmentInfo) -> bool:
-    # An element marked for removal is never placed, so it must not keep the block from
-    # returning early below: the placement code would then have nothing left to place.
-    if element.marked_for_removal:
-        return False
-    if element.type == UsfmUpdateBlockElementType.PARAGRAPH:
-        return alignment_info["paragraph_behavior"] == UpdateUsfmMarkerBehavior.PRESERVE and len(element.tokens) == 1
-    if element.type == UsfmUpdateBlockElementType.STYLE:
-        return alignment_info["style_behavior"] == UpdateUsfmMarkerBehavior.PRESERVE
-    return False
 
 
 class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
@@ -56,7 +38,9 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
             len(elements) == 0
             or alignment_info["alignment"].row_count == 0
             or alignment_info["alignment"].column_count == 0
-            or not any(_is_placeable(e, alignment_info) for e in elements)
+            or not any(
+                e.is_placeable(alignment_info["paragraph_behavior"], alignment_info["style_behavior"]) for e in elements
+            )
         ):
             return block
 
@@ -79,9 +63,8 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
                         elements.pop(i)
             elif not (
                 element.type == UsfmUpdateBlockElementType.EMBED
-                # OTHER elements are never transferred, so they must not count as content here
                 or element.type == UsfmUpdateBlockElementType.OTHER
-                or (element.type == UsfmUpdateBlockElementType.TEXT and len(_element_text(element).strip()) == 0)
+                or (element.type == UsfmUpdateBlockElementType.TEXT and len(element.get_text().strip()) == 0)
             ):
                 eob_empty_paras = False
 
@@ -99,7 +82,7 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
         for element in elements:
             if element.type == UsfmUpdateBlockElementType.TEXT:
                 if element.marked_for_removal:
-                    text = _element_text(element)
+                    text = element.get_text()
                     src_sent += text
 
                     # Track seen tokens
@@ -110,7 +93,7 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
                     if len(text.strip()) > 0:
                         src_tok_idx += 1
                 else:
-                    trg_sent += _element_text(element)
+                    trg_sent += element.get_text()
 
             if element.marked_for_removal or (
                 element.type == UsfmUpdateBlockElementType.PARAGRAPH
@@ -122,8 +105,6 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
             elif element.type in [UsfmUpdateBlockElementType.PARAGRAPH, UsfmUpdateBlockElementType.STYLE]:
                 to_place.append(element)
                 adj_src_toks.append(src_tok_idx)
-            # OTHER elements (attributes, milestones) are intentionally dropped: they are tied to
-            # source text that no longer exists, so there is nowhere to transfer them to.
 
         if len(trg_sent.strip()) == 0:
             return block
@@ -172,8 +153,7 @@ class PlaceMarkersUsfmUpdateBlockHandler(UsfmUpdateBlockHandler):
         to_insert.sort(key=lambda x: x[0])
         to_insert += [(len(trg_sent), element) for element in embed_elements + end_elements]
 
-        # The guard above only predicts that something is placeable. If nothing survived
-        # element collection, leave the block untouched rather than rebuilding its text.
+        # In the case of unclosed markers, to_insert might be empty
         if len(to_insert) == 0:
             return block
