+++ machine/corpora/usfm_update_block_element.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from enum import Enum, auto
 
+from .update_usfm_behavior import UpdateUsfmMarkerBehavior
 from .usfm_token import UsfmToken
 
 
@@ -22,3 +23,17 @@ class UsfmUpdateBlockElement:
         if self.marked_for_removal:
             return []
         return self.tokens.copy()
+
+    def get_text(self) -> str:
+        return "".join(t.to_usfm() for t in self.tokens)
+
+    def is_placeable(
+        self, paragraph_behavior: UpdateUsfmMarkerBehavior, style_behavior: UpdateUsfmMarkerBehavior
+    ) -> bool:
+        if self.marked_for_removal:
+            return False
+        if self.type == UsfmUpdateBlockElementType.PARAGRAPH:
+            return paragraph_behavior == UpdateUsfmMarkerBehavior.PRESERVE and len(self.tokens) == 1
+        if self.type == UsfmUpdateBlockElementType.STYLE:
+            return style_behavior == UpdateUsfmMarkerBehavior.PRESERVE
+        return False
